refactor: Promote operator as a mod for futher refactor
