+++ src/operator/builder.rs
@@ -1,0 +1,122 @@
+// Copyright 2022 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use crate::layers::*;
+use crate::raw::*;
+use crate::*;
+
+/// OperatorBuilder is a typed builder to builder an Operator.
+///
+/// # Notes
+///
+/// OpenDAL uses static dispatch internally and only perform dynamic
+/// dispatch at the outmost type erase layer. OperatorBuilder is the only
+/// public API provided by OpenDAL come with generic parameters.
+///
+/// It's required to call `finish` after the operator built.
+///
+/// # Examples
+///
+/// For users who want to support many services, we can build a helper function like the following:
+///
+/// ```
+/// use std::collections::HashMap;
+///
+/// use opendal::layers::LoggingLayer;
+/// use opendal::layers::RetryLayer;
+/// use opendal::services;
+/// use opendal::Builder;
+/// use opendal::Operator;
+/// use opendal::Result;
+/// use opendal::Scheme;
+///
+/// fn init_service<B: Builder>(cfg: HashMap<String, String>) -> Result<Operator> {
+///     let op = Operator::from_map::<B>(cfg)?
+///         .layer(LoggingLayer::default())
+///         .layer(RetryLayer::new())
+///         .finish();
+///
+///     Ok(op)
+/// }
+///
+/// async fn init(scheme: Scheme, cfg: HashMap<String, String>) -> Result<()> {
+///     let _ = match scheme {
+///         Scheme::S3 => init_service::<services::S3>(cfg)?,
+///         Scheme::Fs => init_service::<services::Fs>(cfg)?,
+///         _ => todo!(),
+///     };
+///
+///     Ok(())
+/// }
+/// ```
+pub struct OperatorBuilder<A: Accessor> {
+    accessor: A,
+}
+
+impl<A: Accessor> OperatorBuilder<A> {
+    /// Create a new operator builder.
+    #[allow(clippy::new_ret_no_self)]
+    pub fn new(accessor: A) -> OperatorBuilder<impl Accessor> {
+        // Make sure error context layer has been attached.
+        OperatorBuilder { accessor }
+            .layer(ErrorContextLayer)
+            .layer(CompleteLayer)
+    }
+
+    /// Create a new layer with static dispatch.
+    ///
+    /// # Notes
+    ///
+    /// `OperatorBuilder::layer()` is using static dispatch which is zero
+    /// cost. `Operator::layer()` is using dynamic dispatch which has a
+    /// bit runtime overhead with an extra vtable lookup and unable to
+    /// inline.
+    ///
+    /// It's always recommended to use `OperatorBuilder::layer()` instead.
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// # use std::sync::Arc;
+    /// # use anyhow::Result;
+    /// use opendal::layers::LoggingLayer;
+    /// use opendal::services::Fs;
+    /// use opendal::Operator;
+    ///
+    /// # #[tokio::main]
+    /// # async fn main() -> Result<()> {
+    /// let op = Operator::create(Fs::default())?
+    ///     .layer(LoggingLayer::default())
+    ///     .finish();
+    /// // All operations will go through the new_layer
+    /// let _ = op.object("test_file").read().await?;
+    /// # Ok(())
+    /// # }
+    /// ```
+    #[must_use]
+    pub fn layer<L: Layer<A>>(self, layer: L) -> OperatorBuilder<L::LayeredAccessor> {
+        OperatorBuilder {
+            accessor: layer.layer(self.accessor),
+        }
+    }
+
+    /// Finish the building to construct an Operator.
+    pub fn finish(self) -> Operator {
+        let ob = self.layer(TypeEraseLayer);
+
+        Operator::from(Arc::new(ob.accessor) as FusedAccessor)
+    }
+}
